refactor: deleted `permissions` pkg
