--- internal/permissions/commons/commons.go
@@ -1,9 +1,0 @@
-package commons
-
-type PermissionLevel string
-
-const (
-	OrgnisationLevelPermission PermissionLevel = "OrgnisationLevelPermission"
-	ProjectLevelPermission     PermissionLevel = "ProjectLevelPermission"
-	EnvironmentLevelPermission PermissionLevel = "EnvironmentLevelPermission"
-)
