--- internal/permissions/default.go
@@ -1,14 +1,0 @@
-package permissions
-
-var instance Service
-
-func SetService(svc Service) {
-	if instance != nil {
-		panic("service already assigned")
-	}
-	instance = svc
-}
-
-func GetService() Service {
-	return instance
-}
