--- internal/permissions/init.go
@@ -1,5 +1,0 @@
-package permissions
-
-func init() {
-	SetService(&DefaultPermissionService{})
-}
