--- internal/permissions/project/service.go
@@ -1,44 +1,0 @@
-package project
-
-import (
-	"errors"
-
-	"github.com/envsecrets/envsecrets/internal/clients"
-	"github.com/envsecrets/envsecrets/internal/context"
-	"github.com/envsecrets/envsecrets/internal/permissions/commons"
-	"github.com/machinebox/graphql"
-)
-
-// Insert new permissions.
-func Insert(ctx context.ServiceContext, client *clients.GQLClient, options *commons.ProjectPermissionsInsertOptions) error {
-
-	req := graphql.NewRequest(`
-	mutation MyMutation($project_id: uuid!, $user_id: uuid!, $permissions: jsonb!) {
-		insert_project_level_permissions(objects: {user_id: $user_id, project_id: $project_id, permissions: $permissions}) {
-		  affected_rows
-		}
-	  }	  
-	`)
-
-	req.Var("project_id", options.ProjectID)
-	req.Var("user_id", options.UserID)
-	permissions, err := options.Permissions.Marshal()
-	if err != nil {
-		return err
-	}
-	req.Var("permissions", string(permissions))
-
-	var response map[string]interface{}
-	if err := client.Do(ctx, req, &response); err != nil {
-		return err
-	}
-
-	returned := response["insert_project_level_permissions"].(map[string]interface{})
-
-	affectedRows := returned["affected_rows"].(float64)
-	if affectedRows == 0 {
-		return errors.New("failed to insert permission")
-	}
-
-	return nil
-}
